feat(ui): refactor floating entry and expose workspace status - Extract ActionButtons component from FloatingEntry - Create useCredentialStatus hook to manage auth state centrally - Improve StatusPanel to display current workspace (Personal/Team) - Refactor `Cred.ts` to better track and expose account ID sources - Update styles for the new button layout and badges
